New approach: use partially-decoded records
For instance, membership in a Pusher channel can be represented as

    type alias Member infoType =
        { id : String
        , info : infoType
        }

where `infoType` is chosen by the application. So when we receive
a `Json.Decode.Value` from Pusher that we know is type `Member infoType`
but we don't know what concrete type `infoType` is, we decode the `id`
field and leave the `info` field as a `Json.Decode.Value`.

The application author then needs to write a decoder of type
`Json.Decode.Decoder infoType`, and for an incoming member value `m`,
call `Pusher.decodeMember decoder m`