2nd restart: memberDecoder, not decodeMember, etc.
So
    memberDecoder : Decoder infoType -> Decoder (Member infoType)

rather than `decodeMember` taking both a decoder and a `Decode.Value`
and returning the result.

This should make it much easier to build up decoders for things like
Presence Channels with multiple events.

As a beginner, I didn't realize, or forgot, that for `type Generic foo`,
`Generic` is still a constructor function.