In messageFor, always use .text if present
We used to ignore the .text field of an ErrorRecord for subscription
errors (because Pusher's errors were longer and developer-centric).  But
we're heading for the ability to customize messages, so our customers
can cherry-pick the probably few cases where the error should be
user-centric (401 and 403 HTTP return codes, for instance). Our users
more informed about what's relevent to their users.