Always return the entire object in .json
We had been trying to extract the `data` field of an incoming object,
and falling back on returning the Json.Value for the whole thing if it
had no `data` field. But that seems harder to explain, and might miss
some important information if the incoming port setup is wonky.